tobit {AER}	R Documentation

Tobit Regression

Description

Fitting and testing tobit regression models for censored data.

Usage

  tobit(formula, left = 0, right = Inf, dist = "gaussian", subset = NULL, data = list(), ...)

Arguments

formula
a symbolic description of a regression model of type y ~ x1 + x2 + ....

left
left limit for the censored dependent variable y. If set to -Inf, y is assumed not be left-censored.

right
right limit for the censored dependent variable y. If set to Inf, the default, y is assumed not be right-censored.

dist
assumed distribution for the dependent variable y. This is passed to survreg, see the respective man page for more details.

subset
a specification of the rows to be used.

data
a data frame containing the variables in the model.

...
further arguments passed to survreg.

Details

The function tobit is a convenience interface to survreg (for survival regression, including censored regression) setting different defaults and providing a more convenient interface for specification of the censoring information.

The default is the classical tobit model (Tobin 1958, Greene 2003) assuming a normal distribution for the dependent variable with left-censoring at 0.

Technically, the formula of type y ~ x1 + x2 + ... passed to tobit is simply transformed into a formula suitable for survreg: This means the dependent variable is first censored and then wrapped into a Surv object containing the censoring information which is subsequently passed to survreg, e.g., Surv(ifelse(y <= 0, 0, y), y > 0, type = "left") ~ x1 + x2 + ... for the default settings.

Value

An object of class "tobit" inheriting from class "survreg".

References

Greene, W.H. (2003). Econometric Analysis, 5th edition. Upper Saddle River, NJ: Prentice Hall.

Tobin, J. (1958). Estimation of Relationships for Limited Dependent Variables. Econometrica, 26, 24–36.

Examples

data("Affairs")

## from Table 22.4 in Greene (2003)
fm.tobit <- tobit(affairs ~ age + yearsmarried + religiousness + occupation + rating, data = Affairs)
fm.tobit2 <- tobit(affairs ~ age + yearsmarried + religiousness + occupation + rating,
  right = 4, data = Affairs)

summary(fm.tobit)
summary(fm.tobit2)

[Package AER version 1.0-0 Index]